feat(imports): Reduced the need for f32 in the system and now use U16 to perform better

+++ examples/pps_cycle.rs
@@ -34,7 +34,7 @@ fn main() {
         let response = ap33772s
             .negotiate_power_delivery(
                 PROGRAMMABLE_POWER_SUPPLY_POWER_DATA_OBJECT,
-                Some(ElectricPotential::new::<millivolt>(voltage as u32)),
+                Some(ElectricPotential::new::<millivolt>(voltage)),
                 OperatingCurrentSelection::_1A,
                 &power_data_objects,
             )
