+++ src/commands/thermal_resistances/thermal_resistance_50.rs
@@ -28,7 +28,7 @@ pub struct ThermalResistance50 {
 impl ThermalResistance50 {
     /// Returns the thermal resistance value in ohms at 50 degrees Celsius.
     pub fn thermal_resistance(&self) -> ElectricalResistance {
-        ElectricalResistance::new::<ohm>(u32::from(self.raw_thermal_resistance()))
+        ElectricalResistance::new::<ohm>(self.raw_thermal_resistance())
     }
 }
 impl_two_byte_read_command!(ThermalResistance50, Command::ThermalResistance50);
