+++ src/commands/thresholds/over_current_protection_threshold.rs
@@ -40,13 +40,13 @@ impl OverCurrentProtectionThreshold {
         u16::from(self.raw_current())
             .checked_mul(Self::CURRENT_RESOLUTION)
             .ok_or(Ap33772sError::ConversionFailed)
-            .map(|scaled_current| ElectricCurrent::new::<milliampere>(u32::from(scaled_current)))
+            .map(|scaled_current| ElectricCurrent::new::<milliampere>(scaled_current))
     }
     pub fn convert_current_to_raw_current(current: ElectricCurrent) -> Result<u8, Ap33772sError> {
         let current_ma = current.get::<milliampere>();
-        let raw_value = current_ma / u32::from(Self::CURRENT_RESOLUTION);
+        let raw_value = current_ma / Self::CURRENT_RESOLUTION;
 
-        if raw_value > u32::from(u8::MAX) {
+        if raw_value > u16::from(u8::MAX) {
             return Err(Ap33772sError::ConversionFailed);
         }
